% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks_cached.R
\name{breaks_cached}
\alias{breaks_cached}
\title{Cached scale break function}
\usage{
breaks_cached(breaks)
}
\arguments{
\item{breaks}{A function that takes the limits as input and returns breaks
as output. See \code{ggplot2::continuous_scale} for details.}
}
\value{
A wrapped breaks function suitable for use with ggplot scales.
}
\description{
This helper caches the output of a breaks function (such as
\code{scales::breaks_width()}). This means that the first time the breaks are
computed with the helper, the resulting breaks vector will be stored.
All subsequent invocations of the helper will return the same stored breaks,
regardless of the limits provided.
}
\details{
In general this is not what you want, since the breaks should change when the
limits change.
}
