% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_node_text_filled.R
\docType{data}
\name{geom_node_text_filled}
\alias{geom_node_text_filled}
\alias{GeomTextFilled}
\title{Annotate nodes with text and a background}
\usage{
geom_node_text_filled(
  mapping = NULL,
  data = NULL,
  position = "identity",
  parse = FALSE,
  check_overlap = FALSE,
  show.legend = NA,
  ...
)
}
\arguments{
\item{mapping, data, position, parse, check_overlap, show.legend, ...}{Arguments passed to the geom that powers \code{ggraph::geom_node_text()}.
Note that the additional arguments of that function such as \code{repel} are
not supported here.}
}
\value{
Layer that draws text and background grobs
}
\description{
This geom behaves like \code{ggraph::geom_node_text()} except that it also inserts
a white background behind the text extending to the left margin. This will
only make sense for a horizontal dendrogram graph layout with the root node
on the left.
}
\details{
This background covers up part of the grid rendered by the coord layer.
The reason that this is done as part of the text instead of as a separate
layer is so that we have access to the rendered dimensions of the text grobs.
}
\keyword{datasets}
