% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylepic-package.R
\docType{package}
\name{phylepic-package}
\alias{phylepic-package}
\title{phylepic: Combined Visualisation of Phylogenetic and Epidemiological Data}
\description{
A collection of utilities and ggplot extensions to assist with visualisations in genomic epidemiology. This includes the phylepic chart, a visual combination of a phylogenetic tree and a matched epidemic curve. The included ggplot expensions such as date axes binned by week are relevant for other applications in epidemiology and beyond.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cidm-ph/phylepic}
  \item Report bugs at \url{https://github.com/cidm-ph/phylepic/issues}
}

}
\author{
\strong{Maintainer}: Carl Suster \email{Carl.Suster@health.nsw.gov.au} (\href{https://orcid.org/0000-0001-7021-9380}{ORCID})

Other contributors:
\itemize{
  \item Western Sydney Local Health District, NSW Health [copyright holder]
}

}
\keyword{internal}
