% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylepic.R
\name{phylepic}
\alias{phylepic}
\title{Combine metadata (a line list) with a phylogenetic tree}
\usage{
phylepic(tree, metadata, name, date)
}
\arguments{
\item{tree}{An object convertible to a \code{tbl_graph}. This will usually be a
"phylo" object, but see \link[tidygraph:tbl_graph]{tidygraph::tbl_graph} for more details.}

\item{metadata}{A data frame.}

\item{name}{Column in \code{metadata} that corresponds to the tree's tip labels
(tidy-eval).}

\item{date}{Column in \code{metadata} that contains the date data (class "Date")
for the tips (tidy-eval).}
}
\value{
An object of class "phylepic".
}
\description{
Some checks are performed to catch issues where the metadata and tree tips
don't match up. Any columns in \code{metadata} that are factors have all levels
that do not appear in the data dropped.
}
\examples{
library(ape)

tree <- read.tree(system.file("enteric.newick", package = "phylepic"))
metadata <- read.csv(
  system.file("enteric_metadata.csv", package = "phylepic")
)
phylepic(tree, metadata, name, as.Date(collection_date))
}
