% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bars.R
\name{plot_bars}
\alias{plot_bars}
\title{Plot metadata bars panel}
\usage{
plot_bars(phylepic, ...)
}
\arguments{
\item{phylepic}{object of class "phylepic".}

\item{...}{scale specifications.}
}
\value{
If \code{phylepic} is specified returns a ggplot, otherwise a function
that when passed a "phylepic" object produces a ggplot for use with
\code{\link[=plot.phylepic]{plot.phylepic()}}.
}
\description{
This uses \code{\link[ggplot2:geom_tile]{ggplot2::geom_tile()}} to produce a grid with a row aligned with
each tip on the tree, and a column for each type of data specified. If no
scales are specified, one is created for each factor column in the metadata
table.
}
\seealso{
Other phylepic plots: 
\code{\link{plot.phylepic}()},
\code{\link{plot_calendar}()},
\code{\link{plot_epicurve}()},
\code{\link{plot_tree}()}
}
\concept{phylepic plots}
