% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week.R
\name{week_breaks}
\alias{week_breaks}
\title{Breaks for week-binning date axes}
\usage{
week_breaks(width = 1L, week_start = getOption("phylepic.week_start"))
}
\arguments{
\item{width}{Number of weeks between breaks (e.g. \code{2} will give a break every
fortnight).}

\item{week_start}{Day the week begins (defaults to Monday).
Can be specified as a case-insensitive English weekday name such as "Monday"
or an integer. Since you generally won't want to mix definitions, it is
more convenient to control this globally with the \code{"phylepic.week_start"}
option, e.g. \code{options(phylepic.week_start = "Monday")}.}
}
\value{
A break function suitable for use in \code{\link[ggplot2:scale_date]{ggplot2::scale_x_date()}} et al.
}
\description{
Breaks for week-binning date axes
}
