% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_calendar.R
\docType{data}
\name{GeomCalendar}
\alias{GeomCalendar}
\alias{geom_calendar}
\title{Specialised tile geometry for calendar plots}
\usage{
geom_calendar(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  linejoin = "mitre",
  label_params = list(),
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping, data, stat, position, linejoin, na.rm, show.legend, inherit.aes, ...}{see \code{\link[ggplot2:geom_tile]{ggplot2::geom_tile()}}.}

\item{label_params}{additional parameters for text labels if present
(see \code{\link[ggplot2:geom_text]{ggplot2::geom_text()}}).}
}
\description{
This geom behaves mostly the same as \code{\link[ggplot2:geom_tile]{ggplot2::geom_tile()}} with a few
additions. Firstly, the \code{label} aesthetic is supported to draw text on top of
the tiles. Secondly, out of bounds values can be drawn as arrows at the edge
of the scale (see details below).
}
\details{
Any \code{x} values that are infinite (i.e. \code{-Inf} or \code{Inf}) would normally be
dropped by ggplot's layers. If any such values survive the stat processing,
they will be drawn by \code{geom_calendar()} as triangles at the respective edges
of the scale. This is intended to work with a scale configures to use
\code{\link[=oob_infinite]{oob_infinite()}} for out of bounds handling.
}
\keyword{datasets}
